chore: refactor move sumo swap to sumo folder
