+++ app/components/sumo/UI/SumoSwaps/components/ActionAlert.js
@@ -2,9 +2,9 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { View, StyleSheet, TouchableOpacity } from 'react-native';
 import MaterialIcon from 'react-native-vector-icons/MaterialIcons';
-import Alert, { AlertType } from '../../../Base/Alert';
-import Text from '../../../Base/Text';
-import { useTheme } from '../../../../util/theme';
+import Alert, { AlertType } from '../../../../Base/Alert';
+import Text from '../../../../Base/Text';
+import { useTheme } from '../../../../../util/theme';
 const AlertTypeKeys = Object.keys(AlertType);
 
 const VERTICAL_DISPLACEMENT = 12;
