+++ app/components/sumo/UI/SumoSwaps/components/AssetSwapButton.js
@@ -2,12 +2,12 @@ import React, { useMemo } from 'react';
 import PropTypes from 'prop-types';
 import { StyleSheet, View } from 'react-native';
 
-import Text from '../../../Base/Text';
-import AssetActionButton from '../../AssetActionButton';
+import Text from '../../../../Base/Text';
+import AssetActionButton from '../../../../UI/AssetActionButton';
 import InfoModal from './InfoModal';
 
-import useModalHandler from '../../../Base/hooks/useModalHandler';
-import { strings } from '../../../../../locales/i18n';
+import useModalHandler from '../../../../Base/hooks/useModalHandler';
+import { strings } from '../../../../../../locales/i18n';
 
 const styles = StyleSheet.create({
   disabledButton: {
