+++ app/components/sumo/UI/SumoSwaps/components/InfoModal.js
@@ -3,9 +3,9 @@ import PropTypes from 'prop-types';
 import { StyleSheet, View, TouchableOpacity, SafeAreaView } from 'react-native';
 import Modal from 'react-native-modal';
 import IonicIcon from 'react-native-vector-icons/Ionicons';
-import Text from '../../../Base/Text';
-import Title from '../../../Base/Title';
-import { useTheme } from '../../../../util/theme';
+import Text from '../../../../Base/Text';
+import Title from '../../../../Base/Title';
+import { useTheme } from '../../../../../util/theme';
 
 const createStyles = (colors, shadows) =>
   StyleSheet.create({
