+++ app/components/sumo/UI/SumoSwaps/components/LoadingAnimation/backgroundShapes.js
