+++ app/components/sumo/UI/SumoSwaps/components/QuotesModal.js
@@ -14,19 +14,19 @@ import Modal from 'react-native-modal';
 import IonicIcon from 'react-native-vector-icons/Ionicons';
 import { connect } from 'react-redux';
 import BigNumber from 'bignumber.js';
-import { strings } from '../../../../../locales/i18n';
+import { strings } from '../../../../../../locales/i18n';
 import {
   fromTokenMinimalUnitString,
   renderFromTokenMinimalUnit,
   renderFromWei,
   toWei,
   weiToFiat,
-} from '../../../../util/number';
+} from '../../../../../util/number';
 import { getQuotesSourceMessage } from '../utils';
-import Text from '../../../Base/Text';
-import Title from '../../../Base/Title';
+import Text from '../../../../Base/Text';
+import Title from '../../../../Base/Title';
 import Ratio from './Ratio';
-import { useTheme } from '../../../../util/theme';
+import { useTheme } from '../../../../../util/theme';
 
 const createStyles = (colors, shadows) =>
   StyleSheet.create({
