+++ app/components/sumo/UI/SumoSwaps/components/TokenIcon.test.js
