+++ app/components/sumo/UI/SumoSwaps/components/TokenImportModal.js
@@ -3,13 +3,13 @@ import PropTypes from 'prop-types';
 import { StyleSheet, SafeAreaView, View } from 'react-native';
 import Modal from 'react-native-modal';
 import FAIcon from 'react-native-vector-icons/FontAwesome5';
-import ModalDragger from '../../../Base/ModalDragger';
-import Text from '../../../Base/Text';
-import Alert, { AlertType } from '../../../Base/Alert';
+import ModalDragger from '../../../../Base/ModalDragger';
+import Text from '../../../../Base/Text';
+import Alert, { AlertType } from '../../../../Base/Alert';
 import TokenIcon from './TokenIcon';
-import StyledButton from '../../StyledButton';
-import { strings } from '../../../../../locales/i18n';
-import { useTheme } from '../../../../util/theme';
+import StyledButton from '../../../../UI/StyledButton';
+import { strings } from '../../../../../../locales/i18n';
+import { useTheme } from '../../../../../util/theme';
 
 const createStyles = (colors) =>
   StyleSheet.create({
